+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Moved Sanitizer Provider interfaces to lucid-support
+
 ## v0.1.0 (2022-09-07)
 * Built initial Sanitizer structure
 * Added core Processor and Constraint types